app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.9.2